+++ Sources/UI.cpp
@@ -39,9 +39,9 @@ UI::GameOver(sf::RenderWindow* window)
   sf::Text gameOver(policeUI);
   gameOver.setString("GAME OVER");
   gameOver.setPosition(
-    { window->getSize().x / 2.0f -
+    { (float)window->getSize().x / 2.0f -
         ((gameOver.getCharacterSize() * gameOver.getString().getSize()) / 2.0f),
-      window->getSize().y / 2.0f });
+      (float)window->getSize().y / 2.0f });
   window->draw(gameOver);
 }
 void
@@ -50,8 +50,8 @@ UI::Score(sf::RenderWindow* window)
   sf::Text textScore(policeUI);
   textScore.setString("Score : " + std::to_string(Player::player->score));
   textScore.setPosition(
-    { window->getSize().x - (80.0f + (textScore.getCharacterSize() *
-                                      textScore.getString().getSize())),
+    { (float)window->getSize().x - (80.0f + (textScore.getCharacterSize() *
+                                             textScore.getString().getSize())),
       50.0f });
   window->draw(textScore);
 }
